Reimplement proguard rule gen in KSP
This reorganizes proguard rule generation to run in its own KSP SymbolProcessor. This allows it to be reused by both moshi-sealed KSP and moshi-ir, while also solving K2 support for the latter and avoiding Kotlin IC issues.

For moshi-sealed KSP, it just depends on the processor and should run it automatically with no changes.

For moshi-ir, proguard rule gen in the gradle extension will now seek to configure and apply the KSP plugin. This requires the KSP plugin to be on the buildscript classpath now in order to enable it.